requirements.txt
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
pymongo==4.6.1
python-dateutil==2.8.2
six==1.16.0
Werkzeug==3.0.1
zipp==3.17.0
